refactor: BetAmount 클래스 패키지 이동
