+++ src/main/java/model/judgement/ResultStatus.java
@@ -1,5 +1,7 @@
 package model.judgement;
 
+import model.paticipant.BetAmount;
+
 public enum ResultStatus {
 
     WIN {
