+++ src/test/java/model/betting/BetAmountTest.java
@@ -3,8 +3,8 @@ package model.betting;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import model.judgement.BetAmount;
 import model.judgement.Profit;
+import model.paticipant.BetAmount;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
